create utility for streaming updates
